Restructure fabric network module's folders
